reconstruct crypto/
